[issues-309] - Add a tooling module for Observability related tasks, refactor the micrometer module accordingly
