+++ micrometer/src/test/resources/otel-collector-config.yaml
@@ -22,7 +22,7 @@ exporters:
   prometheus:
     endpoint: "0.0.0.0:49152"
   otlp:
-    endpoint: 0.0.0.0:4317
+    endpoint: 0.0.0.0:4217
     tls:
       insecure: true
