+++ microprofile-fault-tolerance/src/test/java/org/jboss/eap/qe/microprofile/fault/tolerance/util/MicroProfileTelemetryServerSetup.java
@@ -1,0 +1,4 @@
+package org.jboss.eap.qe.microprofile.fault.tolerance.util;
+
+public class MicroProfileTelemetryServerSetup {
+}
