+++ pom.xml
@@ -27,6 +27,7 @@
         <module>microprofile-lra</module>
         <module>microprofile-open-api</module>
         <module>micrometer</module>
+        <module>tooling-observability</module>
     </modules>
 
     <properties>
@@ -49,7 +50,7 @@
         <version.io.rest-assured>5.0.1</version.io.rest-assured>
         <version.jakarta.servlet.jakarta.servlet-api>5.0.0</version.jakarta.servlet.jakarta.servlet-api>
         <version.junit>4.13.1</version.junit>
-        <version.org.eclipse.microprofile>6.1</version.org.eclipse.microprofile>
+        <version.org.eclipse.microprofile>7.0</version.org.eclipse.microprofile>
         <version.org.eclipse.microprofile.lra>2.0</version.org.eclipse.microprofile.lra>
         <version.io.micrometer>1.12.4</version.io.micrometer>
         <version.org.jboss.arquillian>1.7.2.Final</version.org.jboss.arquillian>
@@ -251,6 +252,11 @@
             <dependency>
                 <groupId>org.jboss.eap.qe</groupId>
                 <artifactId>tooling-docker</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.jboss.eap.qe</groupId>
+                <artifactId>tooling-observability</artifactId>
                 <version>${project.version}</version>
                 <scope>test</scope>
             </dependency>
