+++ tooling-observability/src/main/java/org/jboss/eap/qe/observability/jaeger/model/JaegerLog.java
@@ -2,7 +2,7 @@
  * Copyright The WildFly Authors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.jboss.eap.qe.micrometer.model.jaeger;
+package org.jboss.eap.qe.observability.jaeger.model;
 
 import java.util.List;
